feat: LiveKit server instances are using the up-to-date configmap always
Description:

Livekit server deployment now has a label to point at a corresponding configmap. Whenever there is a reconcile cycle, based on this label, the configmap resource version will be updated in the deployment's pod template annotation. Thus triggering a rollout restart for the pods. This way those pods will be "restarted" if there is a new configuration for them.
To achieve this the livekitmesh_controller.go had to be updated in order to watch for configmap and gateway resources. The gateway resource is updated with its  newly available ip address by the stunner-gw-operator (first nodePort IP, then LB IP). There is no need for a logic to fetch the auth server from time to time to watch for new Ice configs. The gateway resource will signal us when to fetch for a new config.